requirements.txt
python-dotenv==1.0.1
pytest==8.3.5
requests==2.31.0
selenium==4.31.0